requirements.txt
cached-property==2.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
dnspython==2.7.0
et_xmlfile==2.0.0
exchangelib==5.5.0
greenlet==3.1.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.2
oauthlib==3.2.2
observable==0.3.2
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
ringcentral==0.9.1
six==1.16.0
SQLAlchemy==2.0.36
sspilib==0.2.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==12.0
